pyproject.toml
[project]
name = "word-embeddings-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "pandas>=3.0.0",
    "sentence-transformers>=5.2.2",
    "umap-learn>=0.5.11",
]
